New method for stashing unstaged changes
The `git stash` method was fragile, messing up merge state and even
failing when the working tree contains a change conflicting with a
change in the index. In addition, removing untracked files was often
not desirable, since it could interfere with unrelated programs.

This new method is simpler, but no longer removes untracked files.
Therefore, you don't get a totally clean state from which to run the
pre-commit hooks, but it should be good enough for most purposes.